Removed Bazel in favour of a simpler npm+tsc+rollup setup.
- Moved source files under src/
- Removed Bazel config files
- Removed unused scripts and entry.ts file.
- Removed .buildkite config
