--- .buildkite/pipeline.yml
@@ -1,10 +1,0 @@
-# Based on https://github.com/buildkite/bazel-example
-steps:
-  - label: ":bazel: Build"
-    commands:
-      - bazel build //...
-  - label: ":bazel: Test"
-    commands:
-      - bazel test //...
-      - echo success
-
