+++ README.md
@@ -28,13 +28,12 @@ You can alternatively manually install the latest version from GitHub master bra
 
 2. Build and install
 
-   You will need to [install
-   Bazel](https://docs.bazel.build/versions/master/install-ubuntu.html) on your
+   You will need to [install NodeJS](https://nodejs.org) on your
    system to run the build tool. Then, you can run the installation script to
    install to `$HOME/.local/share/gnome-shell/extensions/gSnap@micahosborne`.
-
+   
    ```shell
-   bazel run :install-extension
+   npm run install-extension
    ```
 
 3. Log out and log back in. (If you don't want to log out, you can restart gnome
@@ -155,11 +154,3 @@ gSnap is licensed under the [GPL v2+](https://www.gnu.org/licenses/gpl-2.0.html)
 For debugging, enable debug in preferences, and in terminal run journalctl /usr/bin/gnome-shell -f
 
 ## Enjoy!
-
-#PACKAGING & DISTRIBUTION NOTES
-1. npm run i
-2. find and extract dist.tar.gz file in bazel-bin
-3. run ./submit.sh to create zip file
-4. Zip file is located in the dist folder.
-
-Note: this is horrible process, needs to be made easier, but works for uploading to gnome extensions.
