--- concat.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-out=$1
-shift
-cat "$@" > $out
