--- entry.ts
@@ -1,6 +1,0 @@
-import { enable, disable } from "./app"
-
-// GJS import system
-declare function registerExtension(i:Function, e:Function, d:Function): void;
-
-registerExtension(function() {}, enable, disable);
