+++ package.json
@@ -1,47 +1,17 @@
 {
-    "name": "example",
+    "name": "gsnap",
     "version": "0.0.1",
-    "description": "Demo of bazel rules for angular",
-    "license": "Apache 2.0",
-    "dependencies": {},
+    "description": "Organize windows in customizable snap zones like FancyZones on Windows.",
+    "license": "GPL 2.0",
     "devDependencies": {
-        "@bazel/bazelisk": "latest",
-        "@bazel/buildifier": "latest",
-        "@bazel/ibazel": "latest",
-        "@bazel/jasmine": "^2.0.1",
-        "@bazel/karma": "^2.0.1",
-        "@bazel/protractor": "^2.0.1",
-        "@bazel/rollup": "^2.0.1",
-        "@bazel/typescript": "^2.0.1",
-        "@types/chai": "^4.2.12",
-        "@types/jasmine": "^3.5.12",
-        "@types/jsdom": "^16.2.6",
-        "@types/mocha": "^8.0.0",
-        "@types/yargs": "^16.0.0",
-        "chai": "^4.2.0",
-        "jasmine-core": "^3.6.0",
-        "jasmine": "^3.6.1",
-        "jsdom": "^16.4.0",
-        "karma-chrome-launcher": "^3.1.0",
-        "karma-firefox-launcher": "^1.3.0",
-        "karma-jasmine": "^3.1.0",
-        "karma-requirejs": "^1.1.0",
-        "karma-sourcemap-loader": "^0.3.7",
-        "karma": "^5.0.2",
-        "mocha": "^8.1.1",
-        "requirejs": "^2.3.6",
-        "rollup": "^2.23.0",
-        "typescript": "^3.9.7",
-        "yargs": "^16.0.0"
+        "rollup": "^2.70.1",
+        "rollup-plugin-copy": "^3.4.0",
+        "typescript": "^4.6.3"
     },
     "scripts": {
-        "serve": "ibazel run //src:devserver",
-        "pree2e": "bazel build test/...",
-        "i": "bazel run :install-extension --verbose_failures ",
-        "e2e": "concurrently \"bazel run //src:prodserver\" \"while ! nc -z 127.0.0.1 8080; do sleep 1; done && protractor\" --kill-others --success first",
-        "prebuildifier": "bazel build @com_github_bazelbuild_buildtools//buildifier",
-        "buildifier": "find . -type f \\( -name BUILD -or -name BUILD.bazel \\) ! -path \"./node_modules/*\" | xargs $(bazel info bazel-bin)/external/com_github_bazelbuild_buildtools/buildifier/*/buildifier",
-        "format": "git-clang-format",
-        "precommit": "check-clang-format \"yarn format\""
+        "clean": "rm -rf dist build dist.tar.gz",
+        "build": "tsc && rollup --config rollup.config.js",
+        "install-extension": "npm run build && sh install-extension.sh",
+        "pack": "zip -r dist.zip dist/*"
     }
 }
