+++ src/extension/editor.ts
