+++ src/extension/logging.ts
