+++ src/extension/polyfill.js
