+++ src/extension/prefs_builder.ts
