+++ src/extension/setup.js
