+++ src/extension/utils.ts
@@ -1,0 +1,34 @@
+export function getCurrentPath() : string | null {
+    let stack = new Error().stack?.split('\n');
+    if(!stack)
+        return null;
+
+    let extensionStackLine;
+
+    // Search for an occurrence of an extension stack frame
+    // Start at 1 because 0 is the stack frame of this function
+    for (let i = 1; i < stack.length; i++) {
+        if (stack[i].includes('/gnome-shell/extensions/')) {
+            extensionStackLine = stack[i];
+            break;
+        }
+    }
+    if (!extensionStackLine)
+        return null;
+
+    // The stack line is like:
+    //   init([object Object])@/home/user/data/gnome-shell/extensions/u@u.id/prefs.js:8
+    //
+    // In the case that we're importing from
+    // module scope, the first field is blank:
+    //   @/home/user/data/gnome-shell/extensions/u@u.id/prefs.js:8
+    let match = new RegExp('@(.+):\\d+').exec(extensionStackLine);
+    if (!match)
+        return null;
+
+    // local import, as the module is used from outside the gnome-shell process
+    // as well (not this function though)
+
+    let path = match[1];
+    return path.split(":")[0];
+}
