+++ src/images/launcher_16_activated.png
