+++ src/images/launcher_16_hover.png
