+++ src/metadata.json
@@ -1,11 +1,12 @@
 {
   "uuid": "gSnap@micahosborne",
   "name": "gSnap",
-  "description": "Organize windows in customizable snap zones like FancyZones on windows.",
-  "version": 1.1,
+  "description": "Organize windows in customizable snap zones like FancyZones on Windows.",
+  "version": 1.2,
   "shell-version": [
       "40",
-      "41"
+      "41",
+      "42"
   ],
   "url": "https://github.com/micahosborne/gSnap",
   "settings-schema": "org.gnome.shell.extensions.gsnap",
