+++ src/schemas/org.gnome.shell.extensions.gsnap.gschema.xml
