--- tsconfig-app.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es3",
-    "module": "system",
-    "lib": ["es2019"],
-
-    "strict": false,
-    "downlevelIteration": true,
-  }
-}
