--- tsconfig-tools.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    /* Basic Options */
-    "target": "es5",                          /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017', or 'ESNEXT'. */
-    "lib": ["es6", "dom"],                    /* Specify library files to be included in the compilation:  */
-    "strict": true,                           /* Enable all strict type-checking options. */
-    "downlevelIteration": true,               /* Provide full support for iterables in 'for-of', spread, and destructuring when targeting 'ES5' or 'ES3'. */
-  }
-}
