+++ tsconfig.json
@@ -6,11 +6,7 @@
 
     "strict": true,                        /* Enable all strict type-checking options. */
     "downlevelIteration": true,            /* Provide full support for iterables in 'for-of', spread, and destructuring when targeting 'ES5' or 'ES3'. */
+    "outDir": "build/"
   },
-  "exclude": [
-    "./bazel-bin/",
-    "./bazel-out/",
-    "./bazel-gSnap/",
-    "./bazel-testlogs/"
-  ]
+  "include": [ "src/extension/*.ts" ],
 }
